+++ packages/ak-blanket/CHANGELOG.md
@@ -1,5 +1,35 @@
-<a name="1.3.0"></a>
-# 1.3.0 (2016-08-30)
+<a name="3.0.0"></a>
+# 3.0.0 (2016-08-31)
+
+
+
+<a name="ak-blanket@3.0.0"></a>
+# ak-blanket@3.0.0 (2016-08-31)
+
+
+### Code Refactoring
+
+* **eventing:** Changed the way events are exported ([6ab80d0](https://bitbucket.org/atlassian/atlaskit/commits/6ab80d0))
+
+
+### BREAKING CHANGES
+
+* eventing: event exports changed
+
+
+
+<a name="ak-blanket@2.0.0"></a>
+# ak-blanket@2.0.0 (2016-08-31)
+
+
+### Code Refactoring
+
+* **events:** Change the event name to not contain a prefix ([0c299c2](https://bitbucket.org/atlassian/atlaskit/commits/0c299c2))
+
+
+### BREAKING CHANGES
+
+* events: This changes the event name used from 'ak-blanket-click' to 'activate'